refactor: Update various scripts and modules in DualGPUOptimizer for improved functionality and compatibility. Changes include enhancements to GPU detection, memory management, and UI components. This update aims to streamline the user experience and ensure better performance in managing dual GPU configurations for machine learning workloads.

+++ tests/test_vram_fit.py
@@ -19,7 +19,7 @@ def test_split_layers():
     assert p["quant"] in ("awq", "gguf") and not p["disk"]
     # If gguf with gpu_layers, verify layers are reasonable
     if p["quant"] == "gguf" and "gpu_layers" in p:
-        assert 0 < p["gpu_layers"] < int(33 * 32)
+        assert 0 < p["gpu_layers"] < (33 * 32)
 
 
 def test_ram_offload():
